LPS-56785 No need for Base*Portlet. I don't see that in any other module
